perl API env:pwd support
In case of a protected private key file, the environment variable
"pwd" will be read out to obtain the pass phrase.

The example from the test file is protected with the passphrase
"foobar".

pwd=foobar
export foobar
./test

If you don't set the env variable, openssl will prompt you
for the pass phrase