Make all engine fixtures checked
A checked fixture is run after forking while an unchecked one is run outside
it. It is likely that due to the complex engine structure, the unchecked one
doesn't work. It sometimes segfaults on implicit C++ delete operations inside
SoftHSM, probably for memory allocated outside of the fork or similar. The
checked fixtures make it run massively slower but for now we have to live with
it. #58 just got a lot more interesting... :)